requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# UI 
django-admin-datta==1.0.10

# Tooling
django-dynamic-datatb==1.0.25
django-api-generator==1.0.17
django-debug-toolbar==4.2.0

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0

# DB Layer
# psycopg2-binary
# mysqlclient
psycopg2==2.9.9

# mqtt
paho-mqtt==1.6.1
